Refactor: spar test Utils.hs (no semantic change).
